backend files/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
langchain>=0.2.0
langchain-openai>=0.1.0
langgraph>=0.2.0
pdfplumber>=0.11.0
PyPDF2>=3.0.0
python-docx>=1.1.0
werkzeug>=3.0.0
